add logic for deciding on selection method
The logic is that short menus (that fit within a single screen) are
selected from using inputlist(), while longer menus use a new buffer. So
far only the inputlist() method is available, so it is currently used
for all menus, although a warning message is displayed for menus that
would use the new buffer method.